+++ src/App.js
@@ -1,46 +1,17 @@
-import React, { useState } from "react";
-import "./App.css"; // 전체 스타일 관리
-import Content from "./components/layout/content/Content";
-import RequestManagerSidebar from "./components/layout/sidebar/RequestManagerSidebar";
-import ContractManagerSidebar from "./components/layout/sidebar/ContactManagerSidebar";
+import React from "react";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import RequestManager from "./pages/requestManager/RequestManager";
+
 const App = () => {
-  const [isNavOpen, setNavOpen] = useState(false);
-  const [effectClass, setEffectClass] = useState(1);
-
-  const toggleNav = () => {
-    setNavOpen(!isNavOpen);
-  };
-
   return (
-    <div className="page">
-      {/* 
-      <UserSidebar
-        isNavOpen={isNavOpen}
-        toggleNav={toggleNav}
-        effectClass={effectClass}
-        setEffectClass={setEffectClass}
-      />
-      */}
-      <ContractManagerSidebar
-        isNavOpen={isNavOpen}
-        toggleNav={toggleNav}
-        effectClass={effectClass}
-        setEffectClass={setEffectClass}
-      />
-      {/*
-      <RequestManagerSidebar
-        isNavOpen={isNavOpen}
-        toggleNav={toggleNav}
-        effectClass={effectClass}
-        setEffectClass={setEffectClass}
-      />
-      */}
-      <Content
-        isNavOpen={isNavOpen}
-        toggleNav={toggleNav}
-        effectClass={effectClass}
-      />
-    </div>
+    <Router>
+      <Routes>
+        {/* <Route path="/user" component={User} />
+          <Route path="/contractManager" component={ContractManager} /> */}
+        <Route path="/requestManager" component={RequestManager} />
+        <Route path="/" exact></Route>
+      </Routes>
+    </Router>
   );
 };
