+++ src/pages/user/UserMain.js
@@ -1,13 +1,9 @@
-import React, { useState } from "react";
+import React from "react";
 import "./UserMain.css";
-import { BsTable } from "react-icons/bs";
-
 const UserMain = () => {
   return (
     <div className="headerContainer">
-      <div className="content">
-        <UserMain />
-      </div>
+      <div className="content"></div>
     </div>
   );
 };
